working login, reg, profile, update
