--- src/components/common/UpdateUser/UpdateUser.scss
@@ -1,134 +1,0 @@
-.update-user {
-    height: calc(100vh - 160px);
-    overflow-y: hidden;
-    overflow-x: auto;
-    position: relative;
-
-    .orangeBox {
-        z-index: +1;
-        background-color: rgba(225, 92, 49, 0.9);
-        height: 500px;
-        max-height: 500px;
-        max-width: 950px;
-        border-radius: 16px;
-        position: relative;
-        padding: 20px;
-        margin-left: 100px;
-
-        .header {
-            margin-top: 10px;
-
-            h1 {
-                margin: 0;
-                font-family: 'Bebas Neue';
-                font-size: 50px;
-            }
-        }
-
-        .subHeader {
-            margin-top: 0px;
-
-            h4 {
-                margin: 0;
-            }
-        }
-
-        .pizzaIcon {
-            width: 350px;
-            z-index: +1;
-            position: absolute;
-            top: 0px;
-            right: -100px;
-        }
-
-        .bottomText {
-            margin-top: 16px;
-            align-content: center;
-            justify-content: center;
-
-            h5 {
-                margin: 0;
-                align-items: center;
-                text-align: center;
-                font-weight: 600;
-            }
-        }
-
-        .form {
-            display: flex;
-            flex-direction: row;
-            margin-top: 4vh;
-            justify-content: center;
-            align-content: center;
-            text-align: center;
-            flex-wrap: wrap;
-
-            .form-left {
-                flex: 1;
-                margin-top: 0;
-
-                .form-input {
-                    margin-top: 12px;
-                    margin-bottom: 10px;
-                    margin-right: 40px;
-                }
-            }
-
-            .form-right {
-                flex: 1;
-                margin-top: 0px;
-
-                .form-input {
-                    margin-top: 12px;
-                    margin-bottom: 10px;
-                }
-
-            }
-
-            .update-button,
-            .cancel-button {
-                align-self: center;
-                margin-top: 20px;
-                width: 80%;
-                background-color: white;
-                border-bottom: 2px solid black;
-                color: #E15C31;
-                font-family: 'Montserrat';
-                font-weight: 800;
-                font-size: 16px;
-
-                &:hover {
-                    background-color: black;
-                    border-bottom: 2px solid white;
-                    letter-spacing: 4px;
-                }
-            }
-
-            .link {
-                color: white;
-                font-weight: 400;
-                text-decoration: underline;
-
-                &:hover {
-                    color: black;
-                    letter-spacing: 4px;
-                }
-            }
-        }
-    }
-
-    .update-pizza-bg {
-        z-index: -1;
-        height: 150vh;
-        position: absolute;
-        top: 0;
-        left: 50%;
-        transform: translateX(-50%);
-    }
-
-    @media (max-width: 768px) {
-        .orangeBox {
-            margin: 0 auto;
-        }
-    }
-}
