+++ src/components/pages/Profile/Profile.scss
@@ -1,0 +1,135 @@
+.my-profile {
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    height: calc(100vh - 160px);
+    overflow-y: hidden;
+    overflow-x: hidden;
+
+    .profile-orangeBox {
+        align-items: center;
+        justify-content: center;
+        align-content: center;
+        margin: auto auto;
+
+        background-color: rgba(225, 92, 49, 0.9);
+        width: auto;
+        height: auto;
+        max-width: 550px;
+        border-radius: 16px;
+        padding: 20px;
+
+        overflow-wrap: break-word;
+        
+        .header {
+            margin-top: 20px;
+
+            h1 {
+                margin: 0;
+                font-family: 'Bebas Neue';
+                font-size: 50px;
+            }
+        }
+
+        .subHeader {
+            margin-top: 0px;
+
+            h4 {
+                margin: 0;
+            }
+        }
+
+        .profile-details {
+            justify-content: flex-start;
+            align-items: flex-start;
+            text-align: left;
+            margin-top: 30px;
+
+            h5 {
+                margin-top: 10px;
+                font-weight: 400;
+                text-align: left;
+                font-size: 18px;
+                color: white;
+            }
+
+            .detail-label {
+                font-weight: 600;
+                color: #f8d7cc;
+            }
+
+            .detail-value {
+                margin-left: 30px;
+                font-weight: 400;
+            }
+
+            table {
+                td {
+                    padding: 4px;
+                }
+            }
+        }
+
+        .profile-buttons {
+            margin-top: 20px;
+            margin-bottom: auto;
+            position: relative;
+            align-items: center;
+            text-align: center;
+
+            .edit-button {
+                align-self: center;
+                margin: 10px;
+                width: 50%;
+                background-color: white;
+                border-bottom: 2px solid black;
+                color: #E15C31;
+                font-family: 'Montserrat';
+                font-weight: 800;
+                font-size: 16px;
+
+                &:hover {
+                    background-color: black;
+                    border-bottom: 2px solid white;
+                    letter-spacing: 4px;
+                }
+            }
+        }
+    }
+
+    .pizza-bg {
+        z-index: -1;
+        height: 200vh;
+        position: absolute;
+        top: 10%;
+        left: 50%;
+        transform: translateX(-50%);
+    }
+
+    @media (max-width: 768px) {
+        overflow-y: auto;
+
+        .orangeBox {
+            margin: 0 auto;
+        }
+
+        .detail-label {
+            display: block; 
+            margin-bottom: auto; 
+        }
+
+        .detail-value {
+            display: block;
+            margin-bottom: auto; 
+        }
+
+        table {
+            tr {
+                margin-bottom: 2px;
+            }
+            td {
+                padding: 4px;
+            }
+        }
+    }
+}
