+++ src/components/pages/Profile/UpdateUser/UpdateUser.scss
@@ -1,0 +1,211 @@
+.update-user {
+    height: calc(100vh - 160px);
+    overflow-y: hidden;
+    overflow-x: hidden;
+    position: relative;
+    align-items: center;
+    justify-content: center;
+
+    .orangeBox {
+        z-index: +1;
+        background-color: rgba(225, 92, 49, 0.9);
+        border-radius: 16px;
+        position: relative;
+        padding: 20px;
+        margin: auto auto;
+
+        width: auto;
+        height: auto;
+        //max-height: 500px;
+        max-width: 950px;
+
+        overflow-wrap: break-word;
+
+        .header {
+            margin-top: 10px;
+
+            h1 {
+                margin: 0;
+                font-family: 'Bebas Neue';
+                font-size: 50px;
+            }
+        }
+
+        .subHeader {
+            margin-top: 0px;
+
+            h4 {
+                margin: 0;
+            }
+        }
+
+        .pizzaIcon {
+            width: 350px;
+            z-index: +1;
+            position: absolute;
+            top: 0px;
+            right: -100px;
+        }
+
+        .bottomText {
+            margin-top: 16px;
+            align-content: center;
+            justify-content: center;
+
+            h5 {
+                margin: 0;
+                align-items: center;
+                text-align: center;
+                font-weight: 600;
+            }
+        }
+
+        .form {
+            //display: flex;
+            //flex-direction: row;
+            // margin-top: 20px;
+            // margin-bottom: 20px;
+            justify-content: center;
+            align-content: center;
+            text-align: center;
+            //flex-wrap: wrap;
+
+            // .form-left {
+            //     //flex: 1;
+            //     margin-top: 0;
+
+            //     .form-input {
+            //         margin-top: 12px;
+            //         margin-bottom: 10px;
+            //         margin-right: 40px;
+            //     }
+            // }
+
+            // .form-right {
+            //     //flex: 1;
+            //     margin-top: 0px;
+
+            //     .form-input {
+            //         margin-top: 12px;
+            //         margin-bottom: 10px;
+            //     }
+
+            // }
+
+            .form-input-cp {
+                display: flex;
+                flex-direction: row;
+                justify-content: space-between;
+                align-items: center;
+
+                .form-input {
+                    flex: 1;
+                    margin: 0 5px;
+                    width: 50%;
+                }
+            }
+
+            .button-container {
+                margin-top: 20px;
+                justify-content: space-between;
+                align-items: center;
+
+                .update-button,
+                .cancel-button {
+                    max-width: 40%;
+                    flex: 1;
+                    margin: 0 5px;
+                    background-color: white;
+                    border-bottom: 2px solid black;
+                    color: #E15C31;
+                    font-family: 'Montserrat';
+                    font-weight: 800;
+                    font-size: 16px;
+
+                    &:hover {
+                        background-color: black;
+                        border-bottom: 2px solid white;
+                        letter-spacing: 4px;
+                    }
+                }
+            }
+
+
+            .link {
+                color: white;
+                font-weight: 400;
+                text-decoration: underline;
+
+                &:hover {
+                    color: black;
+                    letter-spacing: 4px;
+                }
+            }
+        }
+    }
+
+    .update-pizza-bg {
+        z-index: -1;
+        height: 150vh;
+        position: absolute;
+        top: 0;
+        left: 50%;
+        transform: translateX(-50%);
+    }
+
+    @media (min-width: 768px) {
+        .update-user {
+            overflow-y: auto;
+        }
+
+        .form {
+            display: flex;
+            flex-direction: row;
+            margin-top: 20px;
+            margin-bottom: 20px;
+
+            .form-left {
+                flex: 1;
+                margin-top: 0;
+
+                .form-input {
+                    margin-top: 12px;
+                    margin-bottom: 10px;
+                    margin-right: 40px;
+                }
+            }
+
+            .form-right {
+                flex: 1;
+                margin-top: 0px;
+
+                .form-input {
+                    margin-top: 12px;
+                    margin-bottom: 10px;
+                }
+
+            }
+        }
+    }
+
+    @media (max-width: 768px) {
+        .orangeBox {
+            margin: 0 auto;
+        }
+
+        .form {
+            display: flex;
+            flex-direction: column;
+
+            .form-left,
+            .form-right {
+                flex: 1;
+                margin: auto;
+
+                .form-input {
+                    margin: auto;
+                }
+            }
+        }
+    }
+}
